+++ src/app/menu/Help.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 import readme from "../../../README.md";
+import packageJson from "../../../package.json";
 import Modal from "../../components/Modal";
 import Button from "../../components/Button";
 import Markdown from "../../components/Markdown";
@@ -8,9 +9,12 @@ import Markdown from "../../components/Markdown";
 export default function Help() {
   return (
     <StyledModal trigger={<Button>Help</Button>}>
-      <h2>Help</h2>
+      <StyledHeader>Help</StyledHeader>
+      <Version>
+        Editor version: <b>{packageJson.version}</b>
+      </Version>
       <Markdown>
-        {readme.split("<!-- HELP END -->")[0]}
+        {readme.split(/<!-- HELP .* -->/)[1]}
       </Markdown>
     </StyledModal>
   );
@@ -18,4 +22,14 @@ export default function Help() {
 
 const StyledModal = styled(Modal)`
   width: 40em;
+`;
+
+const StyledHeader = styled.h2`
+  margin-bottom: 0.5rem;
+`;
+
+const Version = styled.p`
+  font-family: sans-serif;
+  margin-bottom: 0.5rem;
+  font-size: 75%;
 `;